requirements.txt
torch
torchvision
numpy
opencv-python
PyYAML
git+https://github.com/mctorch/mctorch.git
